fastapi_modulo/modulos/intelicoop/legacy_source/backend/django_project/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<23.0
celery>=5.4,<6.0
django-redis>=5.4,<6.0
redis>=5.0,<6.0
